When CuckooGuestError occurs, don't mark task as "reported."
Even though the "errors" key is populated in the results returned by the API,
it would be better to have "status" indicate failure as well. This also
prevents the various Processing modules from running, which only generate
errors when there are no analysis results.

I also had to tweak the web interface to allow users to see the details of the
failure.