implement pipe-based logging server (thanks Accuvant)
Brad Spengler from Accuvant came up with this idea and it's pretty neat.

There are two main advantages over this logging system when compared to
the older TCP based logging. (Note that the old logging system using
sockets is still available for backwards compatibility).

First, and most importantly, this allows processes to log data by just
connecting to a pipe and writing data to it (using CreateFile and
WriteFile, respectively). This is a huge simplification over using
sockets on Windows as sockets are implemented through a complex
user-mode component. Especially on edge cases such as process
termination one might find that sockets are not working anymore,
therefore preventing logging of potentially crucial data.

Secondly, if we integrate a kernel driver in the future, this kernel
driver will be able to use the same logging method, as sockets are not
as trivial to use in the kernel as they are in user-mode.