fix issues with human clicker
Fixes two issues in the human clicker, similar to bugfixes applied by
Accuvant (but not sure if they're the same). Basically we don't want to
return false in the window enumerator callback as that implies we'll be
missing potentially useful child windows with buttons we're interested
in clicking. This was the case when a button was found with a value that
we didn't like or in the case where the button value didn't have a value
that we know.