new pipedispatcher handles multiple messages per pipe handle
Before we used to accept a new pipe handle for each pipe call. This new
handle was then started in a new thread - this is kind of time consuming
and caused issues when lots and lots of messages were coming in, i.e.,
lots of parallel threads were created each handling a single message.

The new design opens, much like the pipe logging, a pipe handle per
process (until it closed by an external force) and reuses this pipe
handle for all pipe messages of the process.