optimize away tmpfs.sh and improve reusability of vmbackup
This commit removes the dependency on tmpfs.sh (and deletes it) and also
fixes some bugs in the process. The vms directory now consists solely of
symbolic links (to either vmmount or vmbackup) and temporary files.

Instead of copying all the Virtual Machine data from vms to vmbackup in
the case of tmpfs usage we always write the Virtual Machine data to the
vmbackup directory. This fixes an issue where running setup.sh on a
working setup would result in bird0.vdi being removed - therefore
bricking the whole setup.

With these changes vmbackup remains truly read-only and vms (and
optionally vmmount) can be generated on-the-fly simply by re-running
setup.sh again.