get rid of database queries in process.py worker processes
Various problems arose when doing database queries from within the
worker processes.

In combination with Distributed Cuckoo queries would sometimes throw
an exception when calling Database().view_task() in the RunProcessing
etc functions on the to-be-processed report, therefore missing out on
valuable information.

Other than that this is probably a step towards the direction of
improving sqlite3 support as sqlite3 is not supposed to be used in a
multi-threaded/multi-process way.