Update "child's ready"-syscalls in apicalls.d
There're two possible ways of how a child may be spawned:
(1) it's just fork()’ed and
(2) it's exec*()’ed.
In the former case the child's first words will always be
these three specific syscalls (in that particular order):
    fork()
    thread_selfid(x, y, z)
    bsdthread_register(a, b, c)
and the last one -- bsdthread_register -- always fails returning -1.

Consider the latter case: the child was exec*()’ed. Now it's basically
a different program, so the loader (dyld in our case) will load its image
from scratch. At some point of loading the image (actually when all the
shared libraries are loaded) dyld calls the following syscall:
stat64("/AppleInternal\0") -- I believe it has something to do with internal
tests at Apple or whatever; but the point here is that we can watch
for this particular syscall in order to determine when our child is ready
for attaching dtrace to it.