init from backup, create any new vms, init from backup again
As described in the comments we first revive any remaining VMs from our
backup, then create new VMs if required to do so (if the VMs were
somehow really bricked), and then re-initialize our VMs.

The second initialization of VMs only really applies when tmpfs support
is enabled. Namely, it creates a new mount which contains all the newly
created VMs. If no VMs were added then this step is not necessary, but
it only takes a minute anyway.