update inject.exe and monitor.dll
The updated inject.exe is able to handle injection into processes which
do not have kernel32 loaded. This is the case for some system processes
such as smss.exe (for more information refer to monitor commit bc174d).

The latest monitor has a lot of tweaks and additions. Namely, properly
tested 64-bit support with Javascript analysis when analyzing Internet
Explorer 8, a whole bunch of signature additions including masking of
superfluous and irrelevant information, and a couple of stability fixes.