apicalls.py: stringify arguments
Since we don't do any calculations on arguments, it's safe to store
them as strings (easy to transfer, don't have to be limited with 8
bytes for bson, etc).

If an argument is an integer (or long), it is converted into a hex
string because, hell, hex rules the world.

Should fix #24