separate virustotal into its own file for dropped file support
Besides the regular analysis of the target sample or URL this allows us
to also do VirusTotal lookups for dropped files and, in the future,
perhaps other resources as well.
Introduces the usage of the requests library, I hope that's fine..

Furthermore features a scan variable. When a resource has not been
scanned yet on VirusTotal this option will push VirusTotal to scan the
resource. Of course this doesn't work with a regular API key (due to
VirusTotal's rate limiting), but it's still a nice-to-have.

Note that one workaround is to use a VirusTotal rate limiting proxy,
which spreads the VirusTotal workload to the point where it is not
reaching the actual limit, but a number close to it. Obviously this is
only useful if you don't process a lot of samples *and* if you
re-process the reports after a while (namely, when the rate limiting
proxy has finished with VirusTotal), so that all of the latest scans are
included in the final report.