various setup.sh updates
Copy the authorized keys file from the current user to the cuckoo user
so we can ssh to cuckoo@ as well.

Now that the monitor branch has been merged into master (as well as a
bunch of other stuff) we don't need to branch to monitor anymore.

Hardcode the postgresql cuckoo password to cuckoo for now.

In order for the open file limit and the grub kernel command-line
changes to take effect we have to reboot the machine, do so.