resolve issues with unhashable dictionaries
In Python dictionaries are not hashable types. And as set()'s only
accept hashable objects this is not really working out for us. We thus
change from the current dictionary approach to tuple's which are
immutable and thus hashable.

An alternative solution would be to introduce hashable dictionaries[1],
but for now this will have to do just fine.

[1]: http://stackoverflow.com/questions/1151658/python-hashable-dicts