improve random timeouts when netlog connects back
Every now and then it would happen that for some reason the initial
socket connection would fail and then, as no timeout was provided, it
could take quite a bit before the connection completed - hanging the
entire analysis. Sometimes three seconds but sometimes up to a minute.

By providing a timeout of 0.1 seconds we actively force the socket to
try to connect, generally connecting successfully after the first,
second, or at most a third try.

An example log file of a situation where the initial connect attempt
failed and no timeout was given looks roughly as follow:

2015-07-25 13:18:57,022 [lib.cuckoo.core.guest] DEBUG: Uploading analyzer to guest (id=cuckoo1, ip=192.168.56.102)
2015-07-25 13:18:57,279 [lib.cuckoo.core.guest] DEBUG: cuckoo1: analyzer started with PID 2192
2015-07-25 13:18:57,280 [lib.cuckoo.core.guest] DEBUG: cuckoo1: waiting for completion
2015-07-25 13:18:58,281 [lib.cuckoo.core.guest] DEBUG: cuckoo1: analysis not completed yet (status=2)
2015-07-25 13:18:59,284 [lib.cuckoo.core.guest] DEBUG: cuckoo1: analysis not completed yet (status=2)
2015-07-25 13:19:00,287 [lib.cuckoo.core.guest] DEBUG: cuckoo1: analysis not completed yet (status=2)
2015-07-25 13:19:01,289 [lib.cuckoo.core.guest] DEBUG: cuckoo1: analysis not completed yet (status=2)
2015-07-25 13:19:02,292 [lib.cuckoo.core.guest] DEBUG: cuckoo1: analysis not completed yet (status=2)

Followed by the following message after the connection was successful.
After this patch the following line should show up pretty much instantly
in every analysis.

2015-07-25 13:20:06,453 [lib.cuckoo.core.resultserver] DEBUG: LogHandler for live analysis.log initialized.