print big integers as string addresses
MongoDB is not able to handle integers bigger than 64-bits which we hit
at this point for some reason, so escape the integers. Thanks to Holger
Unterbrink for reporting this issue.
Actually, kernel addresses bigger than 64-bits don't make sense, so
there's a todo to be done here - properly handling the volatility
output.