do not delete rows that share the same tag #624 #573
It used to be that deleting a task with a tag resulted in deleting all
tasks and machines that shared the same tag. This commit simply gets
rid of the whole cascading stuff with its side-effects. In effect this
means that unused tags will remain in the database, but then again we're
talking about at most a few dozen or hundred tags, so it's not that big
of a deal.