get rid of overlay data for signatures
The main reason for getting rid of overlay data is the fact that
scattering information about signatures around between the signature
itself and a magic json blob is not going to make it easier to debug.
(People unaware of overlay data could be looking at a signature for
hours wondering where a certain attribute is coming from).

Furthermore it doesn't seem that overlay data has ever been used for
actual signatures, nor does the utils/community.py script support it and
thus it's not even usable by signatures, rendering them useless.