signature code changes and simplification
Get rid of the separation between "evented signatures" and "complete
signature list" as we are not going to be supporting old-style
signatures anymore (that is, signatures that worked solely by looping
all of the results on their own rather than chiming in on the generic
implementation provided by the Cuckoo core).

Get rid of the process function as its only purpose was helping out for
old-style signatures.

Re-indent parts of the code as they were put behind a useless if
statement which provided a very minimal performance optimization.