rewrite signature stuff towards the new version and code cleanup
Get rid of functionality that was either unused, outdated,
incompatible, or heavy on performance.

Redefine marks by allowing signatures to place marks on specific calls.
The web interface will then be able to either filter the marks
completely, or to show everything during and around the time of the
marks. The current implementation lacked that functionality as it
depended only on one particular start/end marker.

Removed list signature functionality as it would return an empty list at
this point anyway.

Fixed a couple of wrapper functions (hopefully). E.g., files, regkeys,
mutexes.

Removed check api and check argument call functions as they're both
heavy on performance as well as inaccurate. (Why would you search for a
string in all arguments of a call - one should know which variable to
look for).

Get rid of the various thread references as they are no longer in play.