improve loading of 3rd party plugins
Check whether the module of the plugin matches. It is perfectly possible
that we have multiple "namespaces" including parent/child namespace
relationships. Handle these cases by comparing the module name.

Allow per-namespace class attributes to be set. In the case of Cuckoo
this allows one to easily assign Windows/Android/etc platforms to each
signature in a particular namespace.