various setup.sh improvements for longterm
Introduce a separate function for cloning the Cuckoo Git repository.
Immediately get rid of the default cuckoo1 virtual machine.

Redo the way /home/cuckoo is chmod'd so not to chmod -R the whole
cuckoo home directory (which can take a while in setups that have been
running for a while).

Initialize the first couple of virtual machines at the correct point in
time in longterm mode.

Use a directory with the name longcuckoo instead of cuckoo in longterm
mode.