move apk entry point detection outside of the emulator, code cleanup
Before the APK package and main activity were resolved inside the
emulator. Besides the fact that it is much easier to handle errors in
resolving these features out of an APK on the host, moving this
functionality to the host also allows us to skip on the Androguard
dependency inside the emulator (which was statically included before).