do not exhaust local sockets when no network interface is available
It would be possible that the network interface was not yet available
due to which the socket.create_connection() function throws an exception
right away. Following the infinite loop this would exhaust all the
sockets which in turn resulted in not being able to reach the host and
thus a completely failed analysis.