install extra dependencies to resolve insecureplatformwarning
Since requests version 1.9 a new warning has been introduced, namely
InsecureRequestWarning [1]. This warning is fairly annoying as code that
was working just fine previously (although perhaps partially unsafe -
why else would be getting such error now?) is now throwing around these
SSL warnings. This is yet another reason for the dependency pinning as
resolved for issue #602.

Anyway, as per documentation [2], all that is required is installing a
couple of extra dependencies. The changes introduced by this commit are
the result of adding those three dependencies.

It should be noted that an alternative way of resolving this issue is by
running Python version 2.7.9 or later.

[1]: https://urllib3.readthedocs.org/en/latest/security.html#insecurerequestwarning
[2]: https://urllib3.readthedocs.org/en/latest/security.html#openssl-pyopenssl