for now make the logs believe there is no new agent yet
The new Agent is going to require documentation, a blogpost or two, and
all that before people will want to use it anyway. So for now let's not
seed any confusion by printing debug messages about it.
This does not count for those that use VMCloak 0.3.x, they get the new
Agent automatically.