properly fix bson vs 32/64-bit analysis
The monitor now indicates whether it is doing a 32-bit or 64-bit
analysis. We can use this to properly determine and interpret the
pointers and integer values that are passed along from the monitor.

Thanks to Mikael Keri and Holger Unterbrink for indicating they had
issues with the previous changes to the way bson data was handled.