by default disable screenshot ocr with tesseract
In analyses with plenty of screenshots rich of text it can take
tesseract minutes upon minutes to OCR all screenshots. As we're
currently not taking advantage of the OCR'd text in any signatures, nor
do we really display the information anywhere (outside of the JSON
report), it doesn't make sense to enable it by default at this point.