only represent tracked processes in the webif
In the monitor we already featured a variable which would keep a state
of whether a process is "tracked" or not. This track variable was
initially meant to have Cuckoo not wait for that particular process when
deciding whether an analysis is finished or not. Now we're propagating
this track variable further so we can ignore said processes in the web
interface.

In practice this means that the lsass.exe process, which we hook in
order to obtain TLS Master Secrets during an analysis - using which you
can decrypt SSL/HTTPS traffic from the PCAP, has the "notrack" flag set.
An analysis does not wait for it to finish as it never does.
Furthermore, its output is only useful to assist other features (i.e.,
decrypting SSL traffic) and not so much for human interpretation.

In other words - ignore lsass.exe behavioral data in the web interface.