separate directory for keeping monitoring binaries
After having been overwriting the binaries in analyzer/windows/bin for
too many years now, it's time to start building up a small but useful
database of monitoring binaries in data/monitor/.

When releasing new features or bug fixes which also affect the monitor,
we can now update the data/monitor/latest symbolic link to point to the
Git hash of the required monitor version. Upon start of Cuckoo we can
then check whether the symbolic links target exists and report to the
user the need to upgrade to the latest monitoring binaries if it does
not. (In the end this will only occur when a user manually updates his
or her Cuckoo setup or does a Git clone of the Git repository).

Aside from the improved user experience this will also improve our
ability to bisect the monitor against older versions of it. Sometimes
new features break older features or introduce crashes. Being able to
replay analyses with older monitoring binaries we should be able to
have bisect-like functionality more easily (without race condition
problems that are introduced when overwriting the analyzer/windows/bin
directory while running multiple analyses in parallel).