integration with moloch as a reporting module
Initial integration of Moloch with Cuckoo. Some of this code is based on
the Moloch integration in cuckoo-modified by Will Metcalf.

To be honest, even though this code has been written from scratch based
on my personal interpretation of Moloch, who would be able to tell the
difference - there's really only one way to implement this in Cuckoo ;-)