upgrade django conf to 1.8 and introduce jinja2 templating
The Django templating engine is a pain to use so we introduce
transparent Jinja2 templating support so it can be used where necessary.

Integration could be better if support for global context variables (for
STATIC_URL) was properly handled. Currently getting some exceptions that
I'll save for a rainy day.