get rid of suboptimal task processing scheduler
Using a locking mechanism we're able to avoid race conditions without
requiring a scheduler. Note that the SQL statement used has only been
tested for PostgreSQL and, although it may work for MySQL, it definitely
does not support SQLite (but that's not really an issue as a SQLite
database shouldn't be used in parallel anyway).