the road towards native vpn support - its configuration and verification
This commit introduces default configuration with commentary on how to
tweak the network routing and VPN support through the usage of the
utils/rooter.py tool which should be run as root next to Cuckoo (if and
only if network routing support is desired).

It also tries to force a network interface upon each registered virtual
machine as that will be required for routing traffic.