generic way for creating baseline reports
Introduces support for creating and running "baseline tasks". The sole
purpose of these baseline tasks is to get as much information as
possible with regards to the standard setup of a Virtual Machine.

Currently we'll be mostly targeting the filtering of Volatility
results and perhaps some Windows Microsoft network traffic as well.
That is, in a baseline task which, much like a regular task, runs for
about 60 up until 120 seconds in a regular Cuckoo setup, we simply boot
the VM and leave the Cuckoo Agent hanging - we're not actually sending
the Cuckoo Analyzer.

In return Cuckoo will already capture quite some Volatility results &
network traffic - legitimate traffic that one is supposed to see when
running Windows. Later, when doing an actual analysis of a sample, we'll
give these benign results a lower priority in the web interface.

The advantage of the implemented approach here as compared to
introducing a standalone script for creating baseline reports is that we
can simply create a baseline task later on - perhaps even after updating
a VM - in the web interface to create a new baseline report. The Cuckoo
scheduler will take care of the rest.

Thanks to Bart Mauritz and Joshua Beens for their work on creating
baseline reports for Volatility that inspired me to pick up on this
topic ;-) The next commit will contain more of that work.