baseline processing and representation
Introduces the storage/baseline/ directory for storing per-machine
baseline information. Introduces a baseline processing module that
stores baseline reports in case of a baseline task and processes the
memory results in case of a regular analysis.

The web interface has been configured to show differences between the
baseline report and the analysis report. Namely, as per suggestion by
Bart Mauritz and Joshua Beens, volatility results that disappeared while
running an actual analysis are marked as yellow, while new results are
marked as red.