save on millions upon millions of function calls
Now that we're starting to grow our Signature base, so is the processing
speed for each analysis. The majority of our signatures do not use the
on_call() functionality, yet this function will be called for virtually
every signature for every API call in the behavioral logs.

As an example, let's take an analysis with 10,000 function calls. This
would result in ~300 * 10,000 or 3 million function calls. Not to
mention the performance impact seen when a sample has hundreds of
thousands of function calls.

In this commit we introduce a dictionary which keeps a list of
"interested" signatures for each API call name. As before signatures are
only added to this list when their filters match, but this is the case
for every signature that doesn't use the on_call() event.

So in addition to the older filtering we now add that when a signature's
on_call() raises an NotImplementedError, we remove it from the
"interested" list. That way it will only be called once and not
anymore afterwards.

This tweak has been observed to cut down on the function calls to
call_signature(), which is a wrapper around on_call() and alike, from
more than a dozen million calls to just a few ten thousand calls.
This happens to equal to a few dozen seconds of raw CPU performance.

One small note about the implementation. The reversed() is required so
to allow us to remove elements (signature that have not implemented
on_call) from the list on-the-go. Doing the removal of elements while
iterating through the list without the reversed() statement would
incorrectly ignore calling the signature right after the deleted
signature (and thus might yield incorrect results - where some
signatures which should have triggered would not trigger).