very very basic score to indicate maliciousness of the analysis
Although this has been a feature request since day one of Cuckoo's
existence, one could argue that Brad Spengler was first to introduce
this feature.

As-is the feature - as it implemented here - is not worth much, but
it should encourage tailoring of Signatures and our rating towards
a useful system.