track and report dead ip address/port combinations
Detect IP addresses that we couldn't connect to during the analysis and
mark them as "dead hosts". Its detection is pretty much what you'd see
in Wireshark as "TCP Retransmission" - hosts that we're not able to
connect to and to which the guest tries to connect a few times in a row.

For legitimate applications this shouldn't happen as they'll want to
keep their hostnames and IP addresses up-and-running because otherwise,
you know, it ruines their business.

Note that we only report about actual dead hosts in the web interface
(and in our Signature) when routing has been enabled for this analysis.
So for people "still" using custom routing this will not trigger.