executive decission to switch to the iexplore monitoring mode
When we analyze a URL we don't Internet Explorer to take a minute to
fully load while appearing to freeze all the time, this simply doesn't
work practically speaking.

In the iexplore monitoring mode we only look at a very small subset of
the functionality provided by the monitor so to load websites close to
as fast as they would be loaded without the iexplore mode. Of course
this subset should still containing all the critical functionality such
as following created sub-processes, but if anything is missing as-is
this can be added later on.

One final note; including monitoring functions related to exploitation
techniques shouldn't fall under the iexplore monitoring mode per se as
we can simply combine the iexplore mode with the to-be-implemented
"exploit" monitoring mode.