rewrite process memory dumping logic
The monitor now informs the Analyzer when a process is about to finish.
This allows the Analyzer to dump the process memory right before it
actually finishes and thus adds support for dumping memory of processes
that terminate before reaching the critical analysis timeout (if that
even occurs in the first).

Furthermore the new dumping logic, which is now embedded in monitor's
inject.exe utility, also dumps metadata about the memory pages such as
their protection, size, etc. (And by separating the functionality into
a separate utility we now also support 32-bit vs 64-bit differences).

And last but not least we also support dumping processes multiple times
just in case this appears useful in the future.