refactor the sniffer auxiliary module to reflect recent changes
Since the integration of per-analysis network routing we have had the
requirement of providing the network interface for each available
machine. This results in not having to define the network interface for
the sniffer anymore as it is already available for each VM/machine.

Furthermore the ResultServer information has been integrated into each
machinery for a while as well, so we can also take that directly. Which
leaves us with some code cleanups.