replace the obsolete dead machine logic by a big warning
This should get rid of the fairly random logic that would deplete all
VMs one by one assuming that for some reason each of them raised a
CuckooDeadMachine exception.
Hopefully this logic wasn't used anymore to begin with and otherwise
surely we can find a better approach or fix the actual issues that
raised the machine exceptions.