reflect near-realtime detection of changes to the guest status
This functionality will allow us to implement a "finish" button in the
Web interface which will prematurely but properly terminate an analysis.
Especially useful if a timeout longer than a few minutes is provided
which, in the future, will be possible again and allows better or at
least longer instrumentation of potentially interesting samples.