nictrace machine option to explicitly have virtualbox dump the pcap
In cases where one VM talks directly to another VM we seem not to be
able to dump the traffic from the gateway if the VMs belong to the same
subnet, which they might in our case.

Do note that dumping traffic through the nictrace functionality is "not
as nice" as the usual tcpdump approach as we're not able to specify BPF
rules to ignore internal Cuckoo traffic. Such is life.