introduce the concept of additional services
Through the new "services" auxiliary module it is possible to boot extra
VMs next to an analysis. These VMs share the same subnet as the analysis
VM and thus are directly attackable by any malicious sample doing so.

Currently only one VM can be run with additional VMs, but with some
extra task tags magic it should be possible to do multiple analyses in
parallel (and then for each VM there should be N service VMs).