regular analyses will not have machine set
The Scheduler will now first try to fetch tasks with a machine set and
only if that fails will it try to fetch a task without a machine. This
is prone to errors with priority handling but for now will have to do,
to be improved later on when we fully rewrite the scheduling anyway.