workaround for not sending off random analyses to service vms
Through this workaround we basically add an extra check on top of the
earlier "if not machinery.availables()" logic. The service VMs are
specific to the services auxiliary module and should not be used for
actual analyses. Therefore we keep track that, in the case of available
machines, at least one machine is not a service VM, and only in that
case do we spawn a new AnalysisManager thread.
This avoids us from spawning a new AnalysisManager thread every second
when service VMs are defined.
And also when we do spawn a new AnalysisManager instance, we make sure
that the selected task is not a service task.