proper fog integration with cuckoo
As per documentation we were using a cronjob to determine when a new
task should be deployed for a machine, i.e., after which reboot the
machine should be reflashed with its known clean image.
Unfortunately this is prone to race conditions where a new task is
already added to the FOG queue while the current restore operation is
not finished yet, causing the machine to be reflashed twice while we
have not been able to do an analysis in-between.
In the case that the whole reflashing part is actually an expensive or
time consuming process, this will quickly start to frustate its users
who merely want to quickly use the physical machine.

Therefore we introduce some great screenscraping technology in order to
properly automate the task download queueing thing with FOG.

Furthermore we will now automatically start machines that are turned off
through the Wake On LAN protocol. Now if the machines have WOL enabled,
that would be great. Perhaps a timeout should be added in case a machine
is actually started through WOL as it might take a few moments to boot.