remove the nose development package
On some systems I'm getting issues similar to [1] and as this is library
supposed to be used for testing we can omit it from the requirements and
only in such testing scenarios manually install it (e.g., through the
Travis CI configuration). It is clear why the error is happening, i.e.,
/usr/local/man is a dead symbolic link, but at this point I'm not sure
why this happens, so might as well just avoid installing it to prevent
others from experiencing the same issue.

[1]: https://github.com/nose-devs/nose/issues/731