start suricata daemon on startup and change unix socket path
Since we last integrated Suricata some changes have come along.
Apparently it is no longer possible to provide an absolute path for the
Unix socket so instead we now go for a simple filename in the /var/run
directory. (Due this directory we now also have to run Suricata as root
as regular users can't create files / unix sockets in this directory, or
is there a proper way around this?)

Also provide and integrate a script for automatically starting Suricata
upon reboot. Note that it is not possible to start/stop Suricata every
time Cuckoo starts/stops as we can't stop the Suricata service properly
after it has been started (aside from kill -9'ing it). This is due to
the fact that the Suricata service script doesn't allow Suricata to run
*only* in Unix mode (and thus we must invoke it ourselves).