install elasticsearch and bare metal packages in setup.sh
Automatically installs ElasticSearch 2.0 as part of the Cuckoo setup.

Also installs the samba-common-bin and ethtool packages which play a
role in the bare metal support of Cuckoo. The Samba utilities contain
the "net rpc" tool for remotely shutting down a Windows computer and the
ethtool allows one to enable Wake-On-LAN support on a local interface.