slightly improved elasticsearch results parsing
There are some issues regarding ElasticSearch of which the correct
solution is currently not clear to me. Good chance there are various
improvements to be made by using more features of the Query DSL.

For now one of the issues that seems to pop up is as follows. If a file
called "HOWTO_RESTORE_FILES.HTML" is indexed it is possible to query
that result by searching for `howto*`, `howto restore *`, and some more
combinations. However, it seems that, while `howto rest*` provides the
results, `howto resto*` and `howto restor*` don't do their job.

A quick workaround for this issue seems to be querying `"howto resto"*`
and alike, however, a more proper solution should be implemented in the
future. Perhaps some members of the community with more ElasticSearch
experience are willing to take their chances at improving our new
searching functionality.

Other than that we introduce a regex-based matching solution that is
better at identifying which records of a matching document are
interesting to our criteria.