use poll(2) when possible avoiding the 1024 fd limit #719
It is not possible to poll file descriptors of 1024 or higher using the
traditional select(2) approach. On platforms where it is available use
poll(2) instead of select(2) to workaround this issue.

Presumably this issue should occur less often on Windows-based systems
as one could only assume that Windows-based Cuckoo users have just one
or maybe two hosts, making it less likely (definitely not impossible
though!) to run out of file descriptors. (Note that poll(2) is not
supported on most if not all Windows versions).