vpncheck utility, verifies whether vpns are still running
Some items are still on the todo list, though. First of all the vpncheck
utility is operating on a list of VPNs returned by the rooter, which
merely checks which tun devices are currently available. Instead, we
should be extracting the VPNs from the Cuckoo configuration.

This same issue applies to the Cuckoo API - it also takes the VPN list
from the rooter. Keeping in mind future changes we should probably put
some effort into storing the configuration in the database so that it
can be manipulated while Cuckoo itself is running. (Actually this was
already implemented in the longterm repository, so time to merge..)

Last but not least we should allow the user to provide a switch to the
vpncheck utility so that it automatically restarts any broken VPNs. In
that scenario we should also take into account that some VPNs may simply
die after a while and will not come back to life.