.wsf analysis support
Recently a blogpost [1] was released covering a "new" technique used
in-the-wild. Rather than embedding a Javascript .js file, the .wsf
technique allows one to use the JScript.Encode feature to encode the
actual Javascript payload further.

As part of our static analysis we provide the decoded Javascript.
The implementation of this work is a port of scrdec18.c [2], a
JScript.Encode decoder implemented in C.

[1]: https://medium.com/@networksecurity/ransomware-delivered-using-wsf-windows-script-file-551a6589da47
[2]: https://gist.githubusercontent.com/bcse/1834878/raw/7483fb72abbb32aa69b853fdcc9f6f72e7568677/scrdec18.c