simple dns server script (initial code by Mark Schloesser)
Provides a simple-to-use DNS server which should do the job for simple
IPv4 DNS requests. Comes with the option to provide a hardcoded IP
address in the case a domain does not exist (so in case a C2 is down, at
the very least the first payload will be sent over the network).

Based on netsink.py by Mark Schloesser [1].

[1]: https://github.com/rep/cuckoo-contrib/blob/master/netsink/netsink.py