further refine the rooter.py script
Although it removes the cross-compatibility with CentOS this change adds
an error message that users could follow on CentOS-based systems.
Note that the CentOS (or Ubuntu/Debian) support was broken due to the
"service" binary not being an actual file, usually; something that
rooter.py looks for during startup. Anyway, should be fine now.